requirements.txt
streamlit==1.31.0
transformers==4.37.0
torch==2.1.2
pillow==10.2.0
pandas==2.2.0
accelerate==0.26.1
